+++ projects/tests/graphql.spec.ts
@@ -814,7 +814,7 @@ describe('graphql', () => {
           "id": 1,
           "thePriority": 1,
           "title": "task a",
-          "userOnServer": undefined,
+          "userOnServer": "",
         },
         Task {
           "category": null,
@@ -825,7 +825,7 @@ describe('graphql', () => {
           "id": 3,
           "thePriority": 1,
           "title": "task c",
-          "userOnServer": undefined,
+          "userOnServer": "",
         },
       ]
     `)
@@ -866,7 +866,7 @@ describe('graphql', () => {
           "id": 3,
           "thePriority": 1,
           "title": "task c",
-          "userOnServer": undefined,
+          "userOnServer": "",
         },
       ]
     `)
@@ -905,7 +905,7 @@ describe('graphql', () => {
           "id": 1,
           "thePriority": 1,
           "title": "testing",
-          "userOnServer": undefined,
+          "userOnServer": "",
         },
       ]
     `)
@@ -940,7 +940,7 @@ describe('graphql', () => {
           "id": 1,
           "thePriority": 1,
           "title": "testing",
-          "userOnServer": undefined,
+          "userOnServer": "",
         },
       ]
     `)